tools/lsa/pyproject.toml
[project]
name = "lsa"
version = "0.1.0"
description = "Legacy Script Archaeologist - analyze legacy script snapshots"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pdfminer.six>=20221105",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
lsa = "lsa.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lsa"]

[tool.hatch.build.targets.wheel.sources]
"lsa" = "lsa"

[tool.hatch.build]
include = [
    "lsa/**/*.py",
    "lsa/**/*.yaml",
]

[dependency-groups]
dev = [
    "pdfminer-six>=20260107",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
